Expose explicit storage key envelope operations
