+++ crates/walletkit-core/src/storage/tests_utils.rs
@@ -165,8 +165,13 @@ impl InMemoryStorageProvider {
     pub fn open_store(
         &self,
     ) -> Result<crate::storage::CredentialStore, crate::storage::StorageError> {
-        let keys = crate::storage::StorageKeys::from_provider(self, 1000)?;
-        crate::storage::CredentialStore::new(self.paths(), Arc::new(keys))
+        let keys = crate::storage::open_or_create_storage_keys(
+            self.paths(),
+            self.keystore(),
+            self.blob_store(),
+            1000,
+        )?;
+        crate::storage::CredentialStore::new(self.paths(), keys)
     }
 
     pub fn new(root: impl AsRef<Path>) -> Self {
